refactor(tests): extract the published daemon fixture
Lift PublishedDaemonFixture out of PublishedInstanceHealthPluginTests so a
second published-host test class can boot a daemon without depending on the
plugin-load tests. Behaviour is unchanged: the class moves verbatim, taking
the StartupTimeout default it already used with it.
